backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
werkzeug
python-dotenv